Homescreen now displays username
